Formatting and linking changes
