--- webscrape/requirements.txt
@@ -1,4 +1,0 @@
-beautifulsoup4
-requests
-schedule
-time
